# Optimise the FlatIndexBuilder save process.