--- a/libursa/FlatIndexBuilder.h
+++ b/libursa/FlatIndexBuilder.h
@@ -11,6 +11,7 @@ class FlatIndexBuilder : public IndexBuilder {
     // Contains uint64_t values in the following format:
     // [trigram: 24 bits] [file_id: 40 bits]
     std::vector<uint64_t> raw_data;
+    FileId max_fileid_;
 
     void add_trigram(FileId fid, TriGram val);