Saving the index builder is the slowest part of the whole process.
This change makes it much faster.

The slowest part of the save was std::sort on the ngrams. In this change I
rewrite the std::sort to a handwritten radixsort. This accounts for most
of the speedup.

Before (in ms):
1 - 2: 9611
1 - 2: 10055
1 - 2: 9555

After (in ms):
1 - 2: 1871
1 - 2: 1907
1 - 2: 1731

Another thing that was optimised is the process of saving data to disk. It
used to use ofstream::put, I've rewritten it to use a temporary std vector
and only spill the data at the end.

Before (in ms):
3 - 4: 1158
3 - 4: 1136
3 - 4: 1056

After (in ms):
3 - 4: 472
3 - 4: 520
3 - 4: 487

In total, this improves the index dump time significantly:

Before (in ms):
index(100, 1000000): 11797
index(100, 1000000): 12222
index(100, 1000000): 11612

After (in ms):
index(100, 1000000): 3285
index(100, 1000000): 3510
index(100, 1000000): 3166

This does not exhaust all the ways to optimise the indexes - for example,
vectorising the radixsort may make it even faster (though it sounds tricky).